Port rlights.h and shaders_basic_lighting.
I wanted to add lighting to a zig/raylib project, and hit some ABI
compatibility issues with rlights.h. The C is pretty simple, so I just
ported it, and figured it would be useful to have here in case someone
needs the same.

I think it might be useful in the main library, but in raylib it is just
an example, so I followed suit.