fix zig fetch command in README
All URLs within build.zig.zon files must point to archives that never
change.  However, the zig fetch command in the README.md adds a URL
that points to the `devel` git branch whose content changes whenever
the `devel` branch is updated.

I've updated the README a url that zig will resolve to a SHA before it
writes it to the zon file.