remove webhack_c
I am unsure what the point of this file is, so maybe I should not be
removing it. But the comment says to try removing it to see if things
work fine, and things work fine on my machine when I remove it.

I tested 3 or 4 of the builtin examples using commands like this:
zig build raw_stream -Dtarget=wasm32-emscripten --sysroot ~/.emscripten_cache/sysroot

(see also #134; my sysroot setup is a bit weird. but I don't think
that would affect this at all)